Pipeline visualisations

A Pipeline object is automatically constructed as calls to make_*() are made. It stores the relationships between targets, dependencies, and sources.

See also

Public fields

segments

A list of Segment objects

Methods

Public methods

Pipeline$add_source_segment()

Pipeline$add_recipe_segment()

Pipeline$build()

Pipeline$clean()

Pipeline$annotate()

Pipeline$refresh()

Pipeline$print()

Pipeline$save()

Pipeline$clone()

Method add_source_segment()

Add an edge to edges

Add any nodes in private$edges that are missing from private$nodes into private$nodes

Reconstruct Pipeline edges from Segment edges. Called primarily to update outofdateness

Add a pipeline segment corresponding to a make_with_source() call

Usage

Pipeline$add_source_segment(
  source,
  targets,
  dependencies,
  packages,
  envir,
  force
)

Arguments

source

The path to an R script which makes the targets

targets

A character vector of paths to files

dependencies

A character vector of paths to files which the targets depend on

packages

A character vector of names of packages which targets depend on

envir

The environment in which to execute the source or recipe.

force

A logical determining whether or not execution of the source or recipe will be forced (i.e. happen whether or not the targets are out-of-date)

new_edge

An data.frame constructed with new_edge()

Returns

The SegmentSource added to the Pipeline

Method add_recipe_segment()

Add a pipeline segment corresponding to a make_with_recipe() call

Usage

Pipeline$add_recipe_segment(
  recipe,
  targets,
  dependencies,
  packages,
  envir,
  force
)

Arguments

recipe

A language object which, when evaluated, makes the targets

targets

A character vector of paths to files

dependencies

A character vector of paths to files which the targets depend on

packages

A character vector of names of packages which targets depend on

envir

The environment in which to execute the source or recipe.

force

A logical determining whether or not execution of the source or recipe will be forced (i.e. happen whether or not the targets are out-of-date)

Returns

The SegmentRecipe added to the Pipeline

Method build()

Build all targets

Usage

Pipeline$build(quiet = getOption("makepipe.quiet"))

Arguments

quiet

A logical determining whether or not messages are signaled

Returns

self

Method clean()

Clean all targets

Usage

Pipeline$clean()

Returns

self

Method annotate()

Apply annotations to Pipeline

Usage

Pipeline$annotate(tooltips, labels)

Arguments

tooltips

A named character vector mapping nodes in the Pipeline onto tooltips to display on hover-over.

labels

A named character vector mapping nodes in the Pipeline onto labels to display beside them.

Method refresh()

Refresh Pipeline to check outofdateness

Usage

Pipeline$refresh()

Method print()

Display pipeline

Usage

Pipeline$print(...)

Arguments

...

Arguments (other than nodes and edges) to pass to visNetwork::visNetwork()

Returns

self

Method save()

Save pipeline

Usage

Pipeline$save(file, selfcontained = TRUE, background = "white", ...)

Arguments

file

File to save HTML into

selfcontained

Whether to save the HTML as a single self-contained file (with external resources base64 encoded) or a file with external resources placed in an adjacent directory.

background

Text string giving the html background color of the widget. Defaults to white.

...

Arguments (other than nodes and edges) to pass to visNetwork::visNetwork()

Returns

self

Method clone()

The objects of this class are cloneable with this method.

Usage

Pipeline$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.